+++ backend/src/main/java/ca/metricalsky/winston/entity/Video.java
@@ -1,7 +1,10 @@
 package ca.metricalsky.winston.entity;
 
 import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
@@ -33,4 +36,7 @@ public class Video {
     @UpdateTimestamp
     private OffsetDateTime lastFetchedAt;
 
+    @OneToOne(fetch = FetchType.EAGER)
+    @JoinColumn(name = "id", referencedColumnName = "video_id")
+    private VideoCommentsEntity comments;
 }