+++ backend/src/main/java/ca/metricalsky/winston/mapper/entity/VideoMapper.java
@@ -14,6 +14,7 @@ public interface VideoMapper {
     @Mapping(target = "description", source = "snippet.description")
     @Mapping(target = "thumbnailUrl", source = "snippet.thumbnails.high.url")
     @Mapping(target = "publishedAt", source = "snippet.publishedAt")
+    @Mapping(target = "comments", ignore = true)
     @Mapping(target = "commentsDisabled", ignore = true)
     @Mapping(target = "lastFetchedAt", ignore = true)
     Video fromYouTube(Activity activity);